cargo_auto_template_new_cli
Module bin_cli_functions_mod

Functions to work with CLI binary executable projects.

Binary executables need some standard functions to help to develop them efficiently.

Constants

BLUE

ANSI color

GREEN

ANSI color

RED

ANSI color

RESET

ANSI color

YELLOW

ANSI color

Functions

panic_set_hook

The original Rust report of the panic is ugly for the end user

tracing_init

Initialize tracing to file tmp/logs/cargo_auto_template_new_cli.log