cargo_auto_template_new_cli

Module utils_mod

Functions to work with CLI binary executable projects.

Binary executables need some standard functions to help to develop them efficiently.

Macros

pos 🔒

macro to get source code position to log errors before propagation

Constants

BLUE

ANSI color

GREEN

ANSI color

RED

ANSI color

RESET

ANSI color

YELLOW

ANSI color

Functions

tracing_init

Initialize tracing to file logs/cargo_auto_template_new_cli.log. \